separação de funções do feed para impedir a criação de vários eventos que disparam  essas ações repetidamente
