+++ src/pages/feed/add-post.js
@@ -49,134 +49,4 @@ export const addPosts = async (post) => {
      `;
 
   document.querySelector('#postsList').innerHTML += postTemplate;
-
-  const clearPostList = () => {
-    document.querySelector('.posts-list').innerHTML = '';
-  };
-
-  const insertPostList = (posts) => {
-    clearPostList();
-    posts.forEach((eachPost) => {
-      addPosts(eachPost);
-    });
-  };
-
-  const postsListContainer = document.querySelector('#postsList');
-
-  // função editar post
-  postsListContainer.addEventListener('click', async (e) => {
-    const { target } = e;
-    const editPostButton = target.dataset.edit;
-    const cancelEditionButton = target.dataset.cancel;
-    const saveEditionButton = target.dataset.save;
-    const deleteButton = target.dataset.delete;
-    const likeButton = target.dataset.like;
-
-    // Open edit
-    if (editPostButton) {
-      const editPostContainer = target.parentNode.parentNode.parentNode.querySelector('.edit-container');
-      const userPost = target.parentNode.parentNode.parentNode.querySelector('.user-post');
-
-      editPostContainer.classList.toggle('display-none');
-      userPost.classList.toggle('display-none');
-    }
-    // cancel edit
-    if (cancelEditionButton) {
-      const liElement = target.parentNode.parentNode.parentNode.parentNode;
-      const userPost = liElement.querySelector('.user-post');
-      const editcontainer = liElement.querySelector('.edit-container');
-
-      editcontainer.classList.toggle('display-none');
-      userPost.classList.toggle('display-none');
-    }
-    // save edit
-    if (saveEditionButton) {
-      const liElement = target.parentNode.parentNode.parentNode.parentNode;
-      const textArea = liElement.querySelector('.edit-post-textarea');
-      const newText = textArea.value;
-      const postId = textArea.dataset.text;
-
-      await firebase.editPost(newText, postId);
-      const posts = await firebase.loadPosts();
-      insertPostList(posts);
-    }
-
-    if (deleteButton) {
-      const postId = deleteButton;
-      await firebase.deletePost(postId);
-      document.querySelector(`#post-${postId}`).remove();
-    }
-
-    if (likeButton) {
-      const postId = likeButton;
-      await firebase.likePosts(postId, currentUserId);
-      const posts = await firebase.loadPosts();
-      insertPostList(posts);
-    }
-  });
-
-  // função excluir posts
-  // const deleteButtons = document.querySelectorAll('.delete-btn');
-  // for (const button of deleteButtons) {
-  //   button.addEventListener('click', async (e) => {
-  //     e.preventDefault();
-  //     const postId = e.currentTarget.parentNode.id;
-  //     await firebase.deletePost(postId);
-  //     document.querySelector(`#post-${postId}`).remove();
-  //   });
-  // }
-
-  // função like posts
-  // const likeButton = document.querySelectorAll('#like-btn');
-  // likeButtons.forEach.addEventListener('click', async (e) => {
-  //   e.preventDefault();
-  //   const postId = e.currentTarget.parentNode.id;
-  //   await firebase.likePosts(postId);
-  //   const posts = await firebase.loadPosts();
-  //   insertPostList(posts);
-  // });
-
-  // função excluir posts
-
-  // const deleteButtons = document.querySelectorAll('.delete-btn')
-
-  // for (const button of deleteButtons) {
-  //   button.addEventListener('click', (e) => {
-  //     e.preventDefault()
-
-  //     const postId = post.id
-  //     console.log(postId)
-  //     const target = e.target.parentNode
-  //     console.log(target)
-  //     const targetDataset = target.dataset.delete
-  //     console.log(targetDataset)
-
-  //     if (targetDataset == "delete") {
-  //       const liElement = target.parentNode.parentNode.parentNode.parentNode;
-  //       console.log(liElement)
-  //       console.log('Chegou aqui')
-  //       //const modal = liElement.querySelector('.modal-wrapper');
-  //       // const confirmDelete =
-  //       // const cancelDelete =
-  //       // modal.style.display = "block"
-
-  //       // confirmDelete.addEventListener("click", () => {
-  //       //   deletePost(postId)
-  //       //     .then(() => {
-  //       //       divDelete.style.display = "none"
-  //       //       post.remove()
-  //       //     })
-  //       //     .catch(e => {
-  //       //       console.log("erro")
-  //       //     })
-  //       // })
-  //       // cancelDelete.addEventListener("click", () => {
-  //       //   divDelete.style.display = "none"
-  //       // })
-  //     }
-
-  //     // deletePost(e.currentTarget.parentNode.id)
-  //     // document.querySelector('#postsList').innerHTML = ''
-  //   });
-  // };
 };
